Added ComputePipeline and Refactored Pipeline classes
- Removed Pipeline::cleanup()
- Renamed datatypes from vk::Datatype to VkDatatype
- Moved VkPipelineLayout and VkPipeline deletion to Pipeline parent class
- Changed PipelineType to private

+++ Engine/include/file_reader.h
@@ -17,7 +17,7 @@ namespace core {
 	public:
 		static Mesh* readMeshFile(std::string filename);
 		static Texture* readImageFile(std::string filename, const ColorSpace& colorSpace = COLOR_SPACE_SRGB);
-		static char* readBytes(const std::string& filepath, uint32_t* size);
+		static char* readBytes(const std::string& filepath, size_t* size);
 
 	private:
